requirements.txt
fastapi
uvicorn[standard]
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
